Reorganize API and internals of serde_dhall a bit
