+++ abnf_to_pest/Cargo.toml
@@ -14,5 +14,5 @@ doctest = false
 [dependencies]
 abnf = "0.6.0"
 indexmap = "1.0.2"
-itertools = "0.8.0"
+itertools = "0.9.0"
 pretty = "0.5.2"
