+++ serde_dhall/tests/version_numbers.rs
@@ -10,5 +10,8 @@ fn test_html_root_url() {
 
 #[test]
 fn test_readme_mentions_version() {
-    version_sync::assert_contains_regex!("../README.md", "^#### \\[{version}\\]");
+    version_sync::assert_contains_regex!(
+        "../README.md",
+        "^#### \\[{version}\\]"
+    );
 }
